+++ Assets/Game.cs
@@ -7,18 +7,14 @@ using GameCanvas;
 /// </summary>
 public sealed class Game : GameBase
 {
-    // 変数の宣言
     int sec = 0;
-    GcImage sky;
 
     /// <summary>
     /// 初期化処理
     /// </summary>
     public override void InitGame()
     {
-        // キャンバスの大きさを設定します
         gc.ChangeCanvasSize(720, 1280);
-        GcAssets.TryGetImage("BlueSky.png", out sky);
     }
 
     /// <summary>
@@ -26,7 +22,6 @@ public sealed class Game : GameBase
     /// </summary>
     public override void UpdateGame()
     {
-        // 起動からの経過時間を取得します
         sec = (int)gc.TimeSinceStartup;
     }
 
@@ -35,19 +30,12 @@ public sealed class Game : GameBase
     /// </summary>
     public override void DrawGame()
     {
-        // 画面を白で塗りつぶします
         gc.ClearScreen();
-
-        // 青空の画像を描画します
-        gc.DrawImage(sky, 0, 0);
-
-        // 黒の文字を描画します
+        gc.DrawImage("BlueSky.png", 0, 0);
         gc.SetColor(0, 0, 0);
         gc.SetFontSize(48);
-        gc.SetStringAnchor(GcAnchor.UpperLeft);
         gc.DrawString("この文字と青空の画像が", 40, 160);
         gc.DrawString("見えていれば成功です", 40, 270);
-        gc.SetStringAnchor(GcAnchor.UpperRight);
-        gc.DrawString($"{sec}s", 630, 10);
+        gc.DrawString($"{sec}s", 630, 10, anchor: GcAnchor.UpperRight);
     }
 }